Bootstrap Symfony
To make it easier to start this project a standard Symfony micro
skeleton is provided.

You can either use this setup or start it from scratch.

We will assess your commit messages to determine if they are good, explain your
reasons to change, and are atomic.

Good luck with your test!